requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Task queue
celery[redis]>=5.3.6
redis>=5.0.4

# Data validation
pydantic>=2.7.0

# Cloud storage
boto3>=1.34.0
google-cloud-storage>=2.16.0

# Transcription providers
openai-whisper>=20231117
assemblyai>=0.26.0

# AI / LLM
groq>=0.9.0

# Video / audio processing
# ffmpeg-python is a thin wrapper; the ffmpeg binary must be installed separately
ffmpeg-python>=0.2.0

# Face tracking
mediapipe>=0.10.14
opencv-python-headless>=4.9.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.6
hypothesis>=6.100.0
httpx>=0.27.0          # async test client for FastAPI
